requirements.txt
aiohttp
chromadb
spacy
Flask
Flask-Cors
llama-index
python-dotenv
sentence-transformers
Werkzeug
huggingface-hub
numpy
pandas
python-magic
gunicorn
llama-index==0.10.45.post1
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-ollama
llama-index-llms-openai
llama-index-multi-modal-llms-ollama
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
llamaindex-py-client
seaborn
matplotlib
pdfplumber
